Move renderers into its own module Move test_renderers into its own module Finish implementation of VBarChart Add Scale and VScale renderers
