+++ asciimatics/constants.py
@@ -35,3 +35,8 @@ COLOUR_BLUE = 4
 COLOUR_MAGENTA = 5
 COLOUR_CYAN = 6
 COLOUR_WHITE = 7
+
+# Line drawing style constants
+ASCII_LINE = 0
+SINGLE_LINE = 1
+DOUBLE_LINE = 2
