+++ samples/bars.py
@@ -1,5 +1,6 @@
 #!/usr/bin/env python3
 
+from asciimatics.constants import SINGLE_LINE, DOUBLE_LINE, ASCII_LINE
 from asciimatics.effects import Print
 from asciimatics.exceptions import ResizeScreenError
 from asciimatics.renderers import BarChart, VBarChart, FigletText
@@ -40,8 +41,7 @@ def demo(screen):
                   )
         hchart2 = BarChart(10, 25, [wv(1), wv(3), wv(5), wv(7), wv(9)],
                       colour=Screen.COLOUR_GREEN, axes=BarChart.BOTH, scale=2.0)
-        hchart2.border_lines.set_style(BoxTool.ASCII_LINE)
-        hchart2.axes_lines.set_style(BoxTool.ASCII_LINE)
+        hchart2.line_style = ASCII_LINE
         hchart3 = BarChart(10, 40, [wv(1), wv(2), wv(3), wv(4), wv(5), wv(7), wv(8), wv(9)],
                       colour=[c for c in range(1, 8)], bg=[c for c in range(1, 8)],
                       scale=2.0, axes=BarChart.X_AXIS, intervals=0.5, labels=True, border=False)
@@ -57,7 +57,7 @@ def demo(screen):
                           (100, 252, 252)
                       ],
                       char=">", scale=100.0, labels=True, axes=BarChart.X_AXIS)
-        hchart4.border_lines.set_style(BoxTool.SINGLE_LINE)
+        hchart4.line_style = SINGLE_LINE
 
         # Vertical Charts
         vchart1 = VBarChart(11, 22, [fn2, fn2], char="═",
@@ -68,8 +68,7 @@ def demo(screen):
                   )
         vchart2 = VBarChart(12, 14, [wv(1), wv(3), wv(5), wv(7), wv(9)],
                       colour=Screen.COLOUR_GREEN, axes=BarChart.BOTH, scale=2.0, gap=0)
-        vchart2.border_lines.set_style(BoxTool.ASCII_LINE)
-        vchart2.axes_lines.set_style(BoxTool.ASCII_LINE)
+        vchart2.line_style = ASCII_LINE
         vchart3 = VBarChart(12, 39, [wv(1), wv(2), wv(3), wv(4), wv(5), wv(7), wv(8), wv(9)],
                       colour=[c for c in range(1, 8)], bg=[c for c in range(1, 8)], gap=0,
                       scale=2.0, axes=BarChart.Y_AXIS, intervals=0.5, labels=True, border=False)
@@ -85,7 +84,7 @@ def demo(screen):
                           (100, 252, 252)
                       ],
                       char=">", scale=100.0, labels=True, axes=VBarChart.Y_AXIS)
-        vchart4.border_lines.set_style(BoxTool.SINGLE_LINE)
+        vchart4.line_style = SINGLE_LINE
 
         effects = [
             Print(screen, hchart1, x=1, y=1, transparent=False, speed=2),
@@ -94,9 +93,9 @@ def demo(screen):
             Print(screen, hchart4, x=96, y=2, transparent=False, speed=2),
 
             Print(screen, vchart1, x=1, y=13, transparent=False, speed=2),
-            Print(screen, vchart2, x=25, y=12, transparent=False, speed=2),
-            Print(screen, vchart3, x=41, y=12, transparent=False, speed=2),
-            Print(screen, vchart4, x=82, y=12, transparent=False, speed=2),
+            Print(screen, vchart2, x=30, y=12, transparent=False, speed=2),
+            Print(screen, vchart3, x=52, y=12, transparent=False, speed=2),
+            Print(screen, vchart4, x=101, y=12, transparent=False, speed=2),
         ]
 
     scenes.append(Scene(effects, -1))
