+++ tests/renderers/globe.gif
